feat: close remaining MFEM alignment gaps + workspace clippy zero
Functional additions:
- LOR-AMG factory (build_lor_amg_h1 / _3d)
- STEP/IGES → NURBS mesh pipeline w/ auto edge detection
- DataCollection time-series I/O abstraction
- Cross-process AMG coarse-level offd w/ ghost exchange
- Exodus/CGNS boundary face extraction
- Viscoplastic (Perzyna) return-mapping closure

Quality:
- Workspace cargo clippy -- -D warnings: zero across 13 crates
- Fixed 4 test warnings in cad.rs, curved.rs
- Updated IGA module docs (arbitrary degree supported)
- Added GhostExchange::from_channels() public API

Cleanup:
- Removed PLAN-0704-mfem-alignment.md (all items verified done)

+++ crates/assembly/src/contact_self.rs
@@ -25,7 +25,7 @@ pub fn assemble_self_contact_3d<M: MeshTopology>(
     mesh: &M, cfg: &ContactConfig, u: &[f64], lam: &[f64],
     locator: &TetPointLocator, gap_max: f64,
 ) -> (Vec<f64>, CsrMatrix<f64>) {
-    let n = mesh.n_nodes() as usize; let nd = n*3;
+    let n = mesh.n_nodes(); let nd = n*3;
     let mut rhs = vec![0.0; nd]; let mut coo = CooMatrix::new(nd, nd);
     let pn = cfg.penalty_normal;
     let (pt, mu) = match &cfg.friction {
